chore(IconSkeleton): move story to components
